requirements.txt
opencv-python
facenet-pytorch
torch
torchvision
torchaudio
pillow
